Improve ScrapeSettings documentation and move to settings.py
