+++ backend/scraping/context/__init__.py
@@ -1,10 +1,9 @@
 from .context import (
     ScrapeContext,
-    ScrapeSettings,
     call_with_context,
     ctx,
-    default_settings,
     get_active_context,
     wrap_with_context,
 )
 from .robots_txt import RobotsTxt
+from .settings import ScrapeSettings, default_settings
